+++ korge-core/src/commonMain/kotlin/korlibs/image/font/SystemFont.kt
@@ -21,19 +21,19 @@ class SystemFont constructor(override val name: String, val coroutineContext: Co
 
     val nativeSystemFontProvider get() = nativeSystemFontProvider(coroutineContext)
 
-    override fun getFontMetrics(size: Float, metrics: FontMetrics): FontMetrics =
+    override fun getFontMetrics(size: Double, metrics: FontMetrics): FontMetrics =
         metrics.also { nativeSystemFontProvider.getSystemFontMetrics(this, size, metrics) }
 
-    override fun getGlyphMetrics(size: Float, codePoint: Int, metrics: GlyphMetrics, reader: WStringReader?): GlyphMetrics =
+    override fun getGlyphMetrics(size: Double, codePoint: Int, metrics: GlyphMetrics, reader: WStringReader?): GlyphMetrics =
         metrics.also { nativeSystemFontProvider.getSystemFontGlyphMetrics(this, size, codePoint, metrics, reader) }
 
     override fun getKerning(
-        size: Float,
+        size: Double,
         leftCodePoint: Int,
         rightCodePoint: Int
-    ): Float = nativeSystemFontProvider.getSystemFontKerning(this, size, leftCodePoint, rightCodePoint)
+    ): Double = nativeSystemFontProvider.getSystemFontKerning(this, size, leftCodePoint, rightCodePoint)
 
-    override fun getGlyphPath(size: Float, codePoint: Int, path: GlyphPath, reader: WStringReader?): GlyphPath? {
+    override fun getGlyphPath(size: Double, codePoint: Int, path: GlyphPath, reader: WStringReader?): GlyphPath? {
         return nativeSystemFontProvider.getSystemFontGlyph(this, size, codePoint, path, reader)
     }
