+++ korge/src/commonMain/kotlin/korlibs/graphics/AG.kt
@@ -44,7 +44,7 @@ interface AGFeatures {
 abstract class AG : AGFeatures, Extra by Extra.Mixin() {
     companion object {
         val logger = Logger("AG")
-        const val defaultPixelsPerInch : Float = 96f
+        const val defaultPixelsPerInch : Double = 96.0
     }
 
     val mainFrameBuffer: AGFrameBuffer = AGFrameBuffer(isMain = true)
