+++ korge/src/commonMain/kotlin/korlibs/korge/view/fast/FSprites.kt
@@ -64,7 +64,7 @@ open class FSprites(val maxSize: Int) {
 
     var FSprite.colorMul: RGBA get() = RGBA(icolorMul32[index]) ; set(value) { icolorMul32[index] = value.value }
 
-    var FSprite.angle: Angle get() = radiansf.radians ; set(value) { radiansf = value.radians }
+    var FSprite.angle: Angle get() = radiansf.radians ; set(value) { radiansf = value.radians.toFloat() }
 
     var FSprite.anchorX: Float get() = unpackAnchorComponent(anchorRaw ushr 0) ; set(value) { anchorRaw = packAnchorComponent(value) or (anchorRaw and 0xFFFF.inv()) }
     var FSprite.anchorY: Float get() = unpackAnchorComponent(anchorRaw ushr 16) ; set(value) { anchorRaw = (packAnchorComponent(value) shl 16) or (anchorRaw and 0xFFFF) }
