+++ korge/src/commonMain/kotlin/korlibs/korge/view/filter/PageFilter.kt
@@ -76,7 +76,7 @@ class PageFilter(
         return MarginInt(max(max(abs(hamplitude0), abs(hamplitude1)), abs(hamplitude2)).toIntCeil())
     }
 
-    override fun updateUniforms(ctx: RenderContext, filterScale: Float) {
+    override fun updateUniforms(ctx: RenderContext, filterScale: Double) {
         super.updateUniforms(ctx, filterScale)
 
         ctx[PageUB].push {
